Custom Message & Address copying
Added an event listener that makes the custom message textarea correspond with the custom message checkbox. When you type in a message in the text area. Made it so that when you add info in the billing address form, that the checkbox corresponding to the delivery address, when checked, when copy the information from the billing address and paste it into the correct inputs. If the checkbox to copy the info is unchecked, the copied info in the delivery address will be removed.